requirements.txt
openai-whisper>=20231117
yt-dlp>=2023.12.30
torch>=2.0.0
torchaudio>=2.0.0